+++ engine-io-snap/Changelog.md
@@ -1,3 +1,7 @@
+## 1.0.1
+
+* Updated for the new `ServerAPI` in `engine-io-1.1`.
+
 ## 1.0.0
 
 * Initial release.